pyproject.toml
[project]
name = "studypartner-ai"
version = "0.1.0"
description = "AI study tutor that watches your screen and coaches you in real-time using research-backed techniques"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Chaker Baloch"},
]
keywords = ["ai", "study", "tutor", "gemini", "coaching", "education"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "google-genai>=1.0",
    "google-adk>=0.5",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-ScreenCaptureKit>=10.0",
    "pyobjc-framework-Vision>=10.0",
    "pyobjc-framework-UserNotifications>=10.0",
    "Pillow>=10.0",
    "typer[all]>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
studypartner = "studypartner.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/studypartner"]

[tool.ruff]
target-version = "py311"
line-length = 100
